moved secret type into internal to avoid cycle imports 🙄🤷‍♂️
