+++ internal/secrets.go
@@ -1,0 +1,6 @@
+package internal
+
+type Secret struct {
+	Name  string
+	Value string
+}
