+++ internal/secrets/.snapshots/Test_parser_Parse-path_from_params_is_a_file
@@ -1,6 +1,6 @@
-(map[string][]secrets.Secret) (len=1) {
-  (string) (len=13) "eikc/repo-one": ([]secrets.Secret) (len=1) {
-    (secrets.Secret) {
+(map[string][]internal.Secret) (len=1) {
+  (string) (len=13) "eikc/repo-one": ([]internal.Secret) (len=1) {
+    (internal.Secret) {
       Name: (string) (len=4) "name",
       Value: (string) (len=5) "value"
     }
