+++ internal/secrets/.snapshots/Test_parser_Parse-provided_path_is_a_directory
@@ -1,10 +1,10 @@
-(map[string][]secrets.Secret) (len=1) {
-  (string) (len=9) "eikc/gapp": ([]secrets.Secret) (len=2) {
-    (secrets.Secret) {
+(map[string][]internal.Secret) (len=1) {
+  (string) (len=9) "eikc/gapp": ([]internal.Secret) (len=2) {
+    (internal.Secret) {
       Name: (string) (len=8) "secret 1",
       Value: (string) (len=6) "123321"
     },
-    (secrets.Secret) {
+    (internal.Secret) {
       Name: (string) (len=8) "secret 2",
       Value: (string) (len=8) "secret 2"
     }
