+++ internal/secrets/.snapshots/Test_parser_sortSecrets
@@ -1,24 +1,24 @@
-(map[string][]secrets.Secret) (len=2) {
-  (string) (len=9) "eikc/gapp": ([]secrets.Secret) (len=3) {
-    (secrets.Secret) {
+(map[string][]internal.Secret) (len=2) {
+  (string) (len=9) "eikc/gapp": ([]internal.Secret) (len=3) {
+    (internal.Secret) {
       Name: (string) (len=13) "Test secret 1",
       Value: (string) (len=6) "123321"
     },
-    (secrets.Secret) {
+    (internal.Secret) {
       Name: (string) (len=13) "Test secret 2",
       Value: (string) (len=6) "123321"
     },
-    (secrets.Secret) {
+    (internal.Secret) {
       Name: (string) (len=13) "test secret 3",
       Value: (string) (len=6) "123321"
     }
   },
-  (string) (len=17) "eikc/masscommerce": ([]secrets.Secret) (len=2) {
-    (secrets.Secret) {
+  (string) (len=17) "eikc/masscommerce": ([]internal.Secret) (len=2) {
+    (internal.Secret) {
       Name: (string) (len=13) "Test secret 1",
       Value: (string) (len=6) "123321"
     },
-    (secrets.Secret) {
+    (internal.Secret) {
       Name: (string) (len=13) "Test secret 2",
       Value: (string) (len=6) "123321"
     }
