--- internal/secrets/secrets.go
@@ -1,13 +1,0 @@
-package secrets
-
-type Spinner interface {
-	Start() error
-	Message(msg string)
-	Stop() error
-	Fail() error
-}
-
-type Secret struct {
-	Name  string
-	Value string
-}
